Validation included in external dependency is not executed
When loading a Smithy project in an IDE like VS Code that starts the
LSP. The language server will use smithy-build.json file to pull
dependencies so that your IDE experience is improved when dealing
with Smithy model that include shapes coming fron elsewhere.

It is expected that the validators included in those external jars
are executed so that relevant information can be surfaced in the IDE.

When using the #assembler builder method, make sur you pass in the
list of external dependencies if they're available, otherwise only
shapes from those jars are kept.