Avoid an exception when the buffer is empty
When looking for a token, during `hover` processing, there is
an exception: ArrayIndexOutOfBound. First, print the exception
when unable to find hover info.

Generally, the problem happens because findToken does not check
if the file is empty. I've updated the code so that we return
a Optional token.

I've also replaced some usage of `getShape`/`get` for `expectShape`.